+++ GRF.Tests/GRFTests.cs
@@ -14,22 +14,17 @@ namespace GRF.Tests
             "Data/test200.grf"
         };
 
-        [Test]
-        public void EntryNames_ReturnsEmptyList_BeforeLoadingAFile()
+        public static LoadingMode[] LoadingModes() => new LoadingMode[]
         {
-            var grf = new Grf();
-            var expected = new List<string>();
-
-            var actual = grf.EntryNames;
-
-            Assert.AreEqual( expected, actual );
-        }
-
+            LoadingMode.Immediate,
+            LoadingMode.Deferred
+        };
+        
         [Test]
-        [TestCaseSource( "InputFiles" )]
-        public void EntryNames_ReturnsAllFilesFromTestGrf_AfterLoadingAFile( string inputFile )
+        public void EntryNames_ReturnsAllFilesFromTestGrf_AfterLoadingAFile(
+            [ValueSource( "InputFiles" )] string inputFile,
+            [ValueSource( "LoadingModes" )] LoadingMode mode )
         {
-            var grf = new Grf();
             var expected = new List<string>() {
                 "data\\0_Tex1.bmp",
                 "data\\11001.txt",
@@ -40,7 +35,7 @@ namespace GRF.Tests
                 "data\\monstertalktable.xml",
                 "data\\resnametable.txt",
                 "data\\t2_¹è°æ1-1.bmp" };
-            grf.Load( inputFile );
+            var grf = Grf.FromFile( inputFile, mode );
 
             var actual = grf.EntryNames;
 
@@ -48,10 +43,10 @@ namespace GRF.Tests
         }
 
         [Test]
-        [TestCaseSource( "InputFiles" )]
-        public void Entries_ContainGrfEntriesWithPaths_AfterLoadingAFile( string inputFile )
+        public void Entries_ContainGrfEntriesWithPaths_AfterLoadingAFile( 
+            [ValueSource( "InputFiles" )] string inputFile,
+            [ValueSource( "LoadingModes" )] LoadingMode mode )
         {
-            var grf = new Grf();
             var expectedPaths = new List<string>() {
                 "data\\0_Tex1.bmp",
                 "data\\11001.txt",
@@ -62,20 +57,21 @@ namespace GRF.Tests
                 "data\\monstertalktable.xml",
                 "data\\resnametable.txt",
                 "data\\t2_¹è°æ1-1.bmp" };
-            grf.Load( inputFile );
+            var grf = Grf.FromFile( inputFile, mode );
 
             foreach( var path in expectedPaths )
             {
-                int index = grf.EntryNames.IndexOf(path);
-                Assert.AreEqual( path, grf.Entries[index].Path );
+                var entryFound = grf.Find( path, out GrfEntry entry );
+                Assert.IsTrue( entryFound );
+                Assert.AreEqual( path, entry.Path );
             }
         }
 
         [Test]
-        [TestCaseSource( "InputFiles" )]
-        public void Files_ContainGrfEntriesWithNames_AfterLoadingAFile( string inputFile )
+        public void Files_ContainGrfEntriesWithNames_AfterLoadingAFile( 
+            [ValueSource( "InputFiles" )] string inputFile,
+            [ValueSource( "LoadingModes" )] LoadingMode mode )
         {
-            var grf = new Grf();
             var expectedPathAndName = new List<(string, string)>() {
                 ("data\\0_Tex1.bmp", "0_Tex1.bmp"),
                 ("data\\11001.txt", "11001.txt"),
@@ -86,20 +82,21 @@ namespace GRF.Tests
                 ("data\\monstertalktable.xml", "monstertalktable.xml"),
                 ("data\\resnametable.txt", "resnametable.txt"),
                 ("data\\t2_¹è°æ1-1.bmp", "t2_¹è°æ1-1.bmp") };
-            grf.Load( inputFile );
+            var grf = Grf.FromFile( inputFile, mode );
 
             foreach( var (path, name) in expectedPathAndName )
             {
-                int index = grf.EntryNames.IndexOf(path);
-                Assert.AreEqual(name, grf.Entries[index].Name);
+                var entryFound = grf.Find( path, out GrfEntry entry );
+                Assert.IsTrue( entryFound );
+                Assert.AreEqual( name, entry.Name );
             }
         }
 
         [Test]
-        [TestCaseSource( "InputFiles" )]
-        public void Files_ContainGrfEntriesWithTypes_AfterLoadingAFile( string inputFile )
+        public void Files_ContainGrfEntriesWithTypes_AfterLoadingAFile(
+            [ValueSource( "InputFiles" )] string inputFile,
+            [ValueSource( "LoadingModes" )] LoadingMode mode )
         {
-            var grf = new Grf();
             var expectedPathAndTypes = new List<(string, string)>() {
                 ("data\\0_Tex1.bmp", "bmp"),
                 ("data\\11001.txt", "txt"),
@@ -110,173 +107,78 @@ namespace GRF.Tests
                 ("data\\monstertalktable.xml", "xml"),
                 ("data\\resnametable.txt", "txt"),
                 ("data\\t2_¹è°æ1-1.bmp", "bmp") };
-            grf.Load( inputFile );
+            var grf = Grf.FromFile( inputFile, mode );
 
             foreach( var (path, type) in expectedPathAndTypes )
             {
-                int index = grf.EntryNames.IndexOf(path);
-                Assert.AreEqual(type, grf.Entries[index].Type);
+                var entryFound = grf.Find( path, out GrfEntry entry );
+                Assert.IsTrue( entryFound );
+                Assert.AreEqual( type, entry.Type);
             }
         }
 
         [Test]
-        [TestCaseSource( "InputFiles" )]
-        public void UncompressedSize_ReturnsSameSizeAsExtractedData_AfterLoadingAFile( string inputFile )
+        public void UncompressedSize_ReturnsSameSizeAsExtractedData_AfterLoadingAFile(
+            [ValueSource( "InputFiles" )] string inputFile,
+            [ValueSource( "LoadingModes" )] LoadingMode mode )
         {
-            var grf = new Grf();
-            grf.Load( inputFile );
+            var grf = Grf.FromFile( inputFile, mode );
 
-            var entries = grf.Entries;
-
-            Assert.IsNotEmpty( entries );
-            entries.ForEach(entry => Assert.AreEqual(entry.header.uncompressedSize, entry.GetUncompressedData().Length));
+            Assert.IsTrue( grf.Count != 0 );
+            foreach(var path in grf.EntryNames )
+            {
+                var entryFound = grf.Find( path, out GrfEntry entry );
+                Assert.IsTrue( entryFound );
+                Assert.AreEqual( entry.Size, entry.GetUncompressedData().Length );
+            }
         }
 
         [Test]
-        [TestCaseSource( "InputFiles" )]
-        public void GetUncompressedData_DoesntChangeOriginalDataOnUncompressing_AfterLoadingAFile( string inputFile )
+        public void GetUncompressedData_DoesntChangeOriginalDataOnUncompressing_AfterLoadingAFile(
+            [ValueSource( "InputFiles" )] string inputFile,
+            [ValueSource( "LoadingModes" )] LoadingMode mode )
         {
-            var grf = new Grf();
-            grf.Load( inputFile );
+            var grf = Grf.FromFile( inputFile, mode );
 
-            var entries = grf.Entries;
-
-            Assert.IsNotEmpty( entries );
-            entries.ForEach(entry => Assert.AreEqual(entry.header.uncompressedSize, entry.GetUncompressedData().Length));
+            Assert.IsTrue( grf.Count != 0 );
+            foreach( var path in grf.EntryNames )
+            {
+                var entryFound = grf.Find( path, out GrfEntry entry );
+                Assert.IsTrue( entryFound );
+                Assert.AreEqual( entry.Size, entry.GetUncompressedData().Length );
+                Assert.AreEqual( entry.Size, entry.GetUncompressedData().Length );
+            }
         }
 
         [Test]
-        [TestCaseSource( "InputFiles" )]
-        public void EntryNames_ReturnsEmptyList_AfterUnloadingAPreviouslyLoadedFile( string inputFile )
+        public void EntryCount_ReturnsNine_AfterLoadingAFile(
+            [ValueSource( "InputFiles" )] string inputFile,
+            [ValueSource( "LoadingModes" )] LoadingMode mode )
         {
-            var grf = new Grf();
-            var expected = new List<string>();
-            grf.Load( inputFile );
-            grf.Unload();
-
-            var actual = grf.EntryNames;
-
-            Assert.AreEqual( expected, actual );
-        }
-
-        [Test]
-        public void EntryCount_ReturnsZero_BeforeLoadingAFile()
-        {
-            var grf = new Grf();
-            var expected = 0;
-
-            var actual = grf.EntryCount;
-
-            Assert.AreEqual( expected, actual );
-        }
-
-        [Test]
-        [TestCaseSource( "InputFiles" )]
-        public void EntryCount_ReturnsNine_AfterLoadingAFile( string inputFile )
-        {
-            var grf = new Grf();
             var expected = 9;
-            grf.Load( inputFile );
+            var grf = Grf.FromFile( inputFile, mode );
 
-            var actual = grf.EntryCount;
-
-            Assert.AreEqual( expected, actual );
-        }
-
-        [Test]
-        [TestCaseSource( "InputFiles" )]
-        public void EntryCount_ReturnsZero_AfterUnloadingAPreviouslyLoadedFile( string inputFile )
-        {
-            var grf = new Grf();
-            var expected = 0;
-            grf.Load( inputFile );
-            grf.Unload();
-
-            var actual = grf.EntryCount;
+            var actual = grf.Count;
 
             Assert.AreEqual( expected, actual );
         }
 
         [Test]
-        public void IsLoaded_ReturnsFalse_BeforeLoadingAFile()
+        public void Load_ThrowsDirectoryNotFound_WhenPassingInvalidPath(
+            [ValueSource( "LoadingModes" )] LoadingMode mode )
         {
-            var grf = new Grf();
-            var expected = false;
+            void throwingMethod() { Grf.FromFile( "some/path/file.grf", mode ); }
 
-            var actual = grf.IsLoaded;
-
-            Assert.AreEqual( expected, actual );
+            Assert.Throws<DirectoryNotFoundException>( throwingMethod );
         }
 
         [Test]
-        [TestCaseSource( "InputFiles" )]
-        public void IsLoaded_ReturnsTrue_AfterLoadingAFile( string inputFile )
+        public void Signature_ReturnsMasterOfMagic_AfterLoadingAFile(
+            [ValueSource( "InputFiles" )] string inputFile,
+            [ValueSource( "LoadingModes" )] LoadingMode mode )
         {
-            var grf = new Grf();
-            var expected = true;
-            grf.Load( inputFile );
-
-            var actual = grf.IsLoaded;
-
-            Assert.AreEqual( expected, actual );
-        }
-
-        [Test]
-        [TestCaseSource( "InputFiles" )]
-        public void IsLoaded_ReturnsFalse_AfterUnloadingAPreviouslyLoadedFile( string inputFile )
-        {
-            var grf = new Grf();
-            var expected = false;
-            grf.Load( inputFile );
-            grf.Unload();
-
-            var actual = grf.IsLoaded;
-
-            Assert.AreEqual( expected, actual );
-        }
-
-        [Test]
-        public void Load_ThrowsFileNotFound_WhenPassingInvalidPath()
-        {
-            var grf = new Grf();
-
-            void throwingMethod() { grf.Load( "some/path/file.grf" ); }
-
-            Assert.Throws<FileNotFoundException>( throwingMethod );
-        }
-
-        [Test]
-        public void Signature_ReturnsEmptyString_BeforeLoadingAFile()
-        {
-            var grf = new Grf();
-            var expected = string.Empty;
-
-            var actual = grf.Signature;
-
-            Assert.AreEqual( expected, actual );
-        }
-
-        [Test]
-        [TestCaseSource( "InputFiles" )]
-        public void Signature_ReturnsMasterOfMagic_AfterLoadingAFile( string inputFile )
-        {
-            var grf = new Grf();
             var expected = "Master of Magic";
-            grf.Load( inputFile );
-
-            var actual = grf.Signature;
-
-            Assert.AreEqual( expected, actual );
-        }
-
-        [Test]
-        [TestCaseSource( "InputFiles" )]
-        public void Signature_ReturnsEmptyString_AfterUnloadingAPreviouslyLoadedFile( string inputFile )
-        {
-            var grf = new Grf();
-            var expected = string.Empty;
-            grf.Load( inputFile );
-            grf.Unload();
+            var grf = Grf.FromFile( inputFile, mode );
 
             var actual = grf.Signature;