+++ GRF/GrfCollection.cs
@@ -6,15 +6,19 @@ namespace GRF
 {
     public class GrfCollection
     {
-        private List<Grf> _loadedGrfs = new List<Grf>();
+        private List<Grf> _grfs = new List<Grf>();
 
-        public GrfCollection() { }
-        public GrfCollection( string iniFilePath ) => Load( iniFilePath );
-
-        public void Load( string iniFilePath, string sectionName = "Data" )
+        public static GrfCollection FromFile( string iniFilePath, string sectionName = "Data", LoadingMode loadingMode = LoadingMode.Deferred )
         {
-            _loadedGrfs.Clear();
+            var grfCollection = new GrfCollection();
+            grfCollection.Load( iniFilePath, sectionName, loadingMode );
+            return grfCollection;
+        }
 
+        private GrfCollection() { }
+
+        private void Load( string iniFilePath, string sectionName, LoadingMode loadingMode )
+        {
             var dataIni = new GrfIni( iniFilePath );
             var directory = Path.GetDirectoryName( iniFilePath );
             var grfFiles = dataIni.Values( sectionName );
@@ -22,42 +26,30 @@ namespace GRF
             foreach( var grfFile in grfFiles )
             {
                 var filePath = Path.Combine( directory, grfFile );
-                _loadedGrfs.Add( new Grf( filePath ) );
+                _grfs.Add( Grf.FromFile( filePath, loadingMode ) );
             }
         }
 
-        public void Unload()
+        public bool Find( string entryName, out GrfEntry entry )
         {
-            foreach( var grf in _loadedGrfs)
+            entry = null;
+            foreach( var grf in _grfs )
             {
-                grf.Unload();
-            }
-            _loadedGrfs.Clear();
-        }
-
-        public bool FindEntry( string entryPath, out GrfEntry file )
-        {
-            GrfEntry _file = null;
-            int hashCode = entryPath.GetHashCode();
-
-            file = null;
-            foreach(Grf grf in _loadedGrfs) {
-                _file = grf.Entries.FirstOrDefault(entry => entry.GetHashCode().Equals(hashCode));
-                if (file != null)
+                if( grf.Find( entryName, out entry ) )
                     break;
             }
-            file = _file;
 
-            return (file != null);
+            return !( entry is null );
         }
 
         public List<string> AllFileNames()
         {
             var fileNames = new List<string>();
-            foreach( var grf in _loadedGrfs )
+            foreach( var grf in _grfs )
             {
                 fileNames.AddRange( grf.EntryNames );
             }
+
             return fileNames.Distinct().ToList();
         }
     }